% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_games.R
\name{build_games}
\alias{build_games}
\title{Build candidate replacement for Games wiki page}
\usage{
build_games(update = FALSE)
}
\arguments{
\item{update}{If already downloaded previously, should we download again?}
}
\value{
A character vector with the text for the candidate replacement for Games wiki page.
As a side effect also saves to
\code{file.path(tools::R_user_dir("ppwicker", "cache"), "games.creole")}.
}
\description{
\code{build_games()} builds a candidate replacement for Games piecepack wiki page
}
