% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_games.R
\name{update_games}
\alias{update_games}
\title{Update Games wiki page}
\usage{
update_games(
  update = FALSE,
  minor = FALSE,
  summary = "Update Games page",
  username = getOption("ppwicker_username"),
  password = getOption("ppwicker_password")
)
}
\arguments{
\item{update}{If already downloaded previously, should we download again?}

\item{minor}{Is this a \sQuote{minor} wiki edit?}

\item{summary}{Summary for wiki edit}

\item{username}{Piecepack wiki username}

\item{password}{Piecepack wiki password (optional)}
}
\description{
\code{update_games()} updates piecepack wiki Games page.
}
