% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_game_metadata.R
\name{download_game_metadata}
\alias{download_game_metadata}
\title{Download piecepack wiki game metadata}
\usage{
download_game_metadata(update = FALSE)
}
\arguments{
\item{update}{If already downloaded previously, should we download again?}
}
\value{
Returns a (tibble) data frame of piecepack wiki game metadata.
If no previous download or \code{update = TRUE} as a side effect
we download piecepack wiki game metadata to \code{tools::R_user_dir("ppwicker", "cache")}.
}
\description{
\code{download_game_metadata()} downloads piecepack wiki game metadata.
}
